+++ RELEASE_NOTES.md
@@ -1,2 +1,6 @@
+### 5.0.0.1-beta 
+* Adjust FSharp.Core dependency
+* Make Microsoft.FSharp.Data.TypeProviders report an obsolete error (use FSharp.Data.TypeProviders instead)
+
 ### 5.0.0.0-beta 
 * Initial pre-release of post-4.3.0.0 component